pyproject.toml
[project]
name = "litestar-music-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "black>=25.1.0",
    "litestar[standard]>=2.16.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
